requirements.txt
fastapi[standard]
python-dotenv
ollama
strands-agents